invoiceAppBackend/requirements.txt
asgiref
django-cors-headers
djangorestframework-simplejwt
djangorestframework
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv